reorg: send-amount-row, send-amount-max-button
