+++ ui/app/components/ui/currency-input/currency-input.container.js
@@ -1,14 +1,14 @@
 import { connect } from 'react-redux'
 import CurrencyInput from './currency-input.component'
 import { ETH } from '../../../helpers/constants/common'
-import { getMaxModeOn } from '../../../pages/send/send-content/send-amount-row/amount-max-button/amount-max-button.selectors'
+import { getSendMaxModeState } from '../../../pages/send/send.selectors'
 import { getIsMainnet, preferencesSelector } from '../../../selectors'
 
 const mapStateToProps = (state) => {
   const { metamask: { nativeCurrency, currentCurrency, conversionRate } } = state
   const { showFiatInTestnets } = preferencesSelector(state)
   const isMainnet = getIsMainnet(state)
-  const maxModeOn = getMaxModeOn(state)
+  const maxModeOn = getSendMaxModeState(state)
 
   return {
     nativeCurrency,
