--- ui/app/pages/send/send-content/send-amount-row/amount-max-button/amount-max-button.selectors.js
@@ -1,3 +1,0 @@
-export function getMaxModeOn (state) {
-  return state.metamask.send.maxModeOn
-}
