--- ui/app/pages/send/send-content/send-amount-row/amount-max-button/tests/amount-max-button-selectors.test.js
@@ -1,22 +1,0 @@
-import assert from 'assert'
-import {
-  getMaxModeOn,
-} from '../amount-max-button.selectors.js'
-
-describe('amount-max-button selectors', function () {
-
-  describe('getMaxModeOn()', function () {
-    it('should', function () {
-      const state = {
-        metamask: {
-          send: {
-            maxModeOn: null,
-          },
-        },
-      }
-
-      assert.equal(getMaxModeOn(state), null)
-    })
-  })
-
-})
