--- ui/app/pages/send/send-content/send-amount-row/send-amount-row.selectors.js
@@ -1,3 +1,0 @@
-export function sendAmountIsInError (state) {
-  return Boolean(state.send.errors.amount)
-}
