+++ ui/app/pages/send/tests/send-selectors.test.js
@@ -23,6 +23,7 @@ import {
   getSelectedTokenExchangeRate,
   getSelectedTokenToFiatRate,
   getSendAmount,
+  sendAmountIsInError,
   getSendEditingTransactionId,
   getSendErrors,
   getSendFrom,
@@ -518,6 +519,35 @@ describe('send selectors', function () {
           },
         }
       )
+    })
+  })
+
+  describe('send-amount-row selectors', function () {
+
+    describe('sendAmountIsInError()', function () {
+      it('should return true if send.errors.amount is truthy', function () {
+        const state = {
+          send: {
+            errors: {
+              amount: 'abc',
+            },
+          },
+        }
+
+        assert.equal(sendAmountIsInError(state), true)
+      })
+
+      it('should return false if send.errors.amount is falsy', function () {
+        const state = {
+          send: {
+            errors: {
+              amount: null,
+            },
+          },
+        }
+
+        assert.equal(sendAmountIsInError(state), false)
+      })
     })
   })
 })
